rename to compress_plan.py
